// manual-scripts/shared/src/types.ts
export interface CollectionMetaData {
    name?: string;
    symbol?: string;
    elementCount?: number;
    updateAuthority?: string;
    creator?: string;
    description?: string;
    family?: string;
    image?: string;
}

export interface CollectionAttributeData {
    attributes: CollectionAttribute[];
}

export interface CollectionPriceData {
    volume: CollectionVolume;
    floorPrice: CollectionPrice;
    ceilingPrice: CollectionPrice;
}

export interface CollectionAttribute {
    name?: string;
    key?: string;
    values: CollectionAttributeValue[];
}
// ...
export enum EntityType {
    BidTransaction = 'bidTransaction',
    AskTransaction = 'askTransaction',
    CancelBidTransaction = 'cancelBidTransaction',
    CancelAskTransaction = 'cancelAskTransaction',
    ExecuteSaleTransaction = 'executeSaleTransaction',
    MarketplaceTransactionForNFT = 'marketplaceTransaction4NFT',
    MarketplaceTransactionForWallet = 'marketplaceTransaction4Wallet',

    BidOrderState = 'bidOrderState',
    AskOrderState = 'askOrderState',

    UserRefreshTokenEntity = 'userRefreshToken',
    NonceEntity = 'nonce',

    MetaplexOnChainMetadata = 'MetaplexOnChainMetadata',
    MintAddress = 'mintAddress',
    NFTOffchainMetadata = 'NFToffchainMetadata',
    MetaplexOffchainMetadata = 'MetaplexOffchainMetadata',
    Trigger = 'trigger',
    CensorTag = "censorTag",
    CollectionID = "collectionID",
    UserFlag = "userFlag",
    StaleFlag = "staleFlag",

    CollectionAttributeData = 'collectionAttributeData',
    CollectionMetaData = 'collectionMetaData',
    CollectionPriceData = 'collectionPriceData',

    NFT4Collection = 'NFT4Collection',
    Collection4NFT = 'collection4NFT',
    PlaydustCollection = 'playdustCollection',
    Transaction = 'transaction',

    MarketData4Collection = 'marketData4Collection',
    OffchainMetadata4Collection = 'offchainMetadata4Collection',
    NFTRarity4Collection = 'NFTRarity4Collection',

    NFTRarity = 'NFTRarity',
}
// ...
export class OffChainMetadata {
	name: string;
    symbol: string;
	description: string;
    seller_fee_basis_points: number;
	image?: string;
	animation_url?: string;
    external_url: string;
	attributes?: Attribute[];
    collection?: CollectionOLD;
    properties: Properties;

	constructor(args: {
		name: string;
        symbol: string;
        description: string;
        seller_fee_basis_points: number;
        image?: string;
        animation_url?: string;
        external_url: string;
        attributes?: Attribute[];
        collection?: CollectionOLD;
        properties: Properties;
	}) {
		this.name = args.name;
        this.symbol = args.symbol;
        this.description = args.description ;
        this.seller_fee_basis_points = args.seller_fee_basis_points;
        this.image = args.image;
        this.animation_url = args.animation_url;
        this.external_url = args.external_url;
        this.attributes = args.attributes;
        this.collection = args.collection;
        this.properties = args.properties;
	}
};
type CollectionOLD = {
    name: string;
    family: string
};
type CreatorOLD = {
    address: string;
    share: number;
};
type File = {
    uri: string;
    type: string;
    cdn?: boolean;
};
type Properties = {
    files?: File[];
    category: MetadataCategory;
    creators?: CreatorOLD[];
};
export enum MetadataCategory {
	Audio = 'audio',
	Video = 'video',
	Image = 'image',
	VR = 'vr',
};
export type Attribute = {
	trait_type?: string;
	display_type?: string;
	value: string | number;
};
// ...
export interface Entity {
    globalId: string;
    id?: string;
    type?: EntityType;
    primaryEntity?: string;
    createdAt?: Date;
    updatedAt: Date;
    // FIXME: Use a generic type here
    data?: any;
};
// ...
export interface Transaction {
    /** The transaction message */
    message: Message;
    /** The transaction signatures */
    signatures: string[];
}

interface ByMarketplaceCollectionData<T> {
    byMarketplace?: Map<Marketplace, T>;
    global: T;
};

export interface VolumeData {
    d1: number;
    d7: number;
    d30: number;
    total: number;
};

export type CollectionPrice = ByMarketplaceCollectionData<number>;
export type CollectionVolume = ByMarketplaceCollectionData<VolumeData>;